Necessary changes to make Qeum boot to shell
One issue is edk2 bootload now doesn't generate acpi table, so I add AcpiBoardInfoGuid back.
This is a workaround, and will not be upstreamed.

Another gap is that Smm feature is not enable in Qemu poc code now, so define EMU_VARIABLE_ENABLE=TRUE when build UPL.

To build UniversalPayload and Qemu binary, need to install LLVM and use build_qemu_pol.bat

Signed-off-by: Zhiguang Liu <zhiguang.liu@intel.com>